Update README to use look.png to avoid font loading issues
So it seems like the browser actually reads the svg code and tries
rendering it itself. However, what this means is that if we use a
particular font that isn't on the user's system, it'll end up being
replaced with some default font by the browser based on the user's
system, which makes the logo look ugly.

Therefore, exporting the svg to a png and using the png avoids this
problem.