Stop trying to automatically spin-up mongodb in TestData/mongodb
This brittle and attempting to do this automatically, while nice, is
probably misguided since it is more flexible to let the developer
handle it according to how they want to do it, e.g., maybe they prefer
a docker image.

This should also make getting the test passing on Github actions much easier.